Crate dfw

DFW - Docker Firewall Framework in Rust

For detailed introductions, see the README.

License

DFW is licensed under either of

Apache License, Version 2.0, (http://www.apache.org/licenses/LICENSE-2.0)

MIT license (https://opensource.org/licenses/MIT)

at your option.

Re-exports

pub use process::*;

Modules

errors
Errors, using failure.

nftables
This module abstracts various nftables concepts into native Rust types.

process
This module holds the types related to configuration processing and rule creation.

rule
TODO: write documentation

types
The types in this module make up the structure of the configuration-file(s).

util
Utilities module