Struct dfw::iptables::IPTablesLogger

pub struct IPTablesLogger { /* fields omitted */ }

IPTables implementation which does not interact with the iptables binary and does not modify the rules active on the host. It does keep a log of every action executed.

Methods

impl IPTablesLogger

pub fn new() -> IPTablesLogger

Create a new instance of IPTablesLogger

pub fn logs(&self) -> Vec<(String, Option<String>)>

Get the collected logs.