Struct dfw::iptables::IPTablesRestore

pub struct IPTablesRestore { /* fields omitted */ }

IPTables implementation which tracks the functions called and maps it to the text-format used by iptables-restore. Upon calling IPTables::commit this text is then passed onto the iptables-restore. This will have the following effect:

Any existing rules which are not part of chains created by DFW will be removed on commit! This means this backend will control all tables in their entirety!

Any rules which are part of chains created by DFW will be completely recreated.

The recreation of the rules happens atomically thanks to iptables-restore. This both cuts down on the execution time and on the time where vital rules might be missing.

Note

A multitude of methods in this implementation are marked as "unsupported". This means that the call will fail with DFWError::TraitMethodUnimplemented.

(Every method that is marked as unsupported also has a justification as to why it isn't implemented, although for most the reason is that DFW doesn't require it and thus no effort was made.)

Methods

impl IPTablesRestore

pub fn new(ip_version: IPVersion) -> Result<IPTablesRestore>

Create a new instance of IPTablesRestore

Note

This backend will recreate all tables it touches -- usually nat and filter -- which means any other rules in those tables that are created externally will be overwritten.

If you require custom rules, you can specify them in the types::Initialization type.

pub fn get_rules(&self) -> Vec<String>

Retrieve the current text that would be passed to iptables-restore as a vector of lines.