Struct dfw::iptables::Iptables

pub struct Iptables;

Marker struct to implement iptables as a firewall backend.

Implementations

impl Iptables

pub fn get_rules(
rules: Vec<IptablesRule>,
rule_discriminant: IptablesRuleDiscriminants
) -> Vec<String>

Retrieve the current text that would be passed to iptables-restore as a vector of lines.