Module dfw::iptables::types

The types in this module make up the structure of the iptables backend-specific configurations.

Example

[backend_defaults.initialization.v4]
filter = [
    "-P INPUT DROP",
    "-F INPUT",
]
[backend_defaults.initialization.v6]
nat = [
    "-P PREROUTING DROP",
]

Structs

Defaults

The defaults/configuration for the iptables backend.

Initialization

The initialization section allows you to add custom rules to any table in both iptables and ip6tables.