Module dfw::nftables

This module implements the nftables backend for DFW.

Modules

types

The types in this module make up the structure of the nftables backend-specific configurations.

Structs

Nftables

Marker struct to implement nftables as a firewall backend.

Enums

Family

Representation of nftables table-families.

Hook

Representation of nftables chain hooks.

Type

Representation of nftables chain-types.