Module dfw::nftables::types

The types in this module make up the structure of the nftables backend-specific configurations.

Example

[backend_defaults]
custom_tables = { name = "filter", chains = ["input", "forward"]}

[backend_defaults.initialization]
rules = [
    "add table inet custom",
]

Structs

Defaults

The defaults/configuration for the nftables backend.

Initialization

The initialization section allows you to execute any commands against nftables.

Table

Reference to an nftables table, specifically to the input- and forward-chains within it.