Module dfw::process

This module holds the types related to configuration processing and rule creation.

Structs

ProcessContext
Enclosing struct to manage rule processing.

ProcessingOptions
Options to configure the processing procedure.

Enums

ContainerFilter
Option to filter the containers to be processed

Traits

Process
This trait allows a type to define its own processing rules. It is expected to return a list of rules that can be applied with nft.