Struct dfw::types::ContainerToContainer

pub struct ContainerToContainer {
    pub default_policy: ChainPolicy,
    pub rules: Option<Vec<ContainerToContainerRule>>,
}

The container-to-container section, defining how containers can communicate amongst each other.

Fields
default_policy: ChainPolicy

The default_policy defines the default for when there is not a specific rule.

Filtering traffic within the same bridge

Depending on how your host is configured, traffic whose origin and destination interface are the same bridge is not filtered by the kernel netfilter module. This means that this default policy is not honored for traffic between containers that are on the same Docker network, but only for traffic that traverses two bridges.

If your kernel has the br_netfilter kernel-module available, you can set the sysctl net.bridge.bridge-nf-call-iptables to 1 to have the netfilter-module act on traffic within the same bridge, too. You can set this value temporarily like this:

sysctl net.bridge.bridge-nf-call-iptables=1

To permanently set this configuration, take a look at man sysctl.d and man sysctl.conf.

rules: Option<Vec<ContainerToContainerRule>>

An optional list of rules, see ContainerToContainerRule.

Example

The easiest way to define the rules is using TOMLs arrays of tables:

[container_to_container]
default_policy = "drop"

[[container_to_container.rules]]
# first rule here
[[container_to_container.rules]]
# second rule here