Crate dfw

DFW - Docker Firewall Framework in Rust

For detailed introductions, see the README.

License

DFW is licensed under either of

Apache License, Version 2.0, (http://www.apache.org/licenses/LICENSE-2.0)

MIT license (https://opensource.org/licenses/MIT)

at your option.

Modules

errors
Errors, using failure.

iptables
This module implements the iptables backend for DFW.

nftables
This module implements the nftables backend for DFW.

process
This module holds the types related to configuration processing and rule creation.

types
The types in this module make up the structure of the configuration-file(s).

util
Utilities module

Traits

FirewallBackend
This trait is used to distinguish between different firewall backends.