Module dfw::iptables

This module holds the IPTables compatibility trait, allowing the use of multiple implementations for the IPTables type of the rust-iptables crate.

Structs

IPTablesDummy
IPTables implementation which does not interact with the iptables binary and does not modify the rules active on the host.

IPTablesLogger
IPTables implementation which does not interact with the iptables binary and does not modify the rules active on the host. It does keep a log of every action executed.

IPTablesRestore
IPTables implementation which tracks the functions called and maps it to the text-format used by iptables-restore. Upon calling IPTables::commit this text is then passed onto the iptables-restore. This will have the following effect:

Enums

IPVersion
Enum identifying a IP protocol version. Can be used by IPTables implementations to discern between IPv4 rules and IPv6 rules.

Traits

IPTables
Compatibility trait to generalize the API used by rust-iptables.