Module dfw::nftables

This module abstracts various nftables concepts into native Rust types.

Enums

ChainPolicy
Representation of nftables chain policies.

Family
Represenation of nftables table-families.

Hook
Representation of nftables chain hooks.

RuleVerdict
Representation of nftables rule policies.

Type
Representation of nftables chain-types.

Functions

add_base_chain
Construct nft command for adding a base chain.

add_chain
Construct nft command for adding a base chain.

add_rule
Construct nft command for adding a rule to a chain.

add_table
Construct nft command for adding a table.

delete_table
Construct nft command for deleting a table.

flush_table
Construct nft command for flushing a table.

insert_rule
Construct nft command for inserting a rule into a chain.

set_chain_policy
Construct nft command for setting the policy for a chain.