Trait dfw::process::Process

pub trait Process {
    fn process(&self, ctx: &ProcessContext) -> Result<Option<Vec<String>>>;
}

This trait allows a type to define its own processing rules. It is expected to return a list of rules that can be applied with nft.

Example

struct MyType {
    rules: Vec<String>,
}

impl Process for MyType {
    fn process(&self, ctx: &ProcessContext) -> Result<Option<Vec<String>>, Error> {
        let mut rules = Vec::new();
        for rule in &self.rules {
            rules.push(format!("add rule {}", rule));
        }
        Ok(Some(rules))
    }
}

Required methods

fn process(&self, ctx: &ProcessContext) -> Result<Option<Vec<String>>>

Process the current type within the given ProcessContext, returning zero or more rules to apply with nft.

Loading content...